discord-bot/requirements.txt
discord-py-interactions==5.3.1
requests==2.28.1
